repackage the core engine; added Javadoc
