+++ Game of Life/Generations/task-info.yaml
@@ -188,7 +188,31 @@ files:
         }
     }
   learner_created: false
-- name: src/life/Model.java
+- name: src/test.java
+  visible: true
+  learner_created: true
+- name: src/test.class
+  visible: true
+  learner_created: true
+- name: src/life/View.class
+  visible: true
+  learner_created: true
+- name: src/life/Life.class
+  visible: true
+  learner_created: true
+- name: src/life/Model.class
+  visible: true
+  learner_created: true
+- name: src/life/Life$LOCALE.class
+  visible: true
+  learner_created: true
+- name: src/life/Life$1.class
+  visible: true
+  learner_created: true
+- name: src/life/Life$STATE.class
+  visible: true
+  learner_created: true
+- name: src/life/engine/Model.java
   visible: true
   text: |
     package life;
@@ -203,19 +227,7 @@ files:
         }
     }
   learner_created: true
-- name: src/life/View.java
-  visible: true
-  text: |
-    package life;
-
-    public interface View {
-        public void view();
-    }
-  learner_created: true
-- name: src/test.java
-  visible: true
-  learner_created: true
-- name: src/life/Life.java
+- name: src/life/engine/Life.java
   visible: true
   text: |
     package life;
@@ -254,26 +266,14 @@ files:
         }
     }
   learner_created: true
-- name: src/life/View.class
+- name: src/life/engine/View.java
   visible: true
-  learner_created: true
-- name: src/life/Life.class
-  visible: true
-  learner_created: true
-- name: src/life/Model.class
-  visible: true
-  learner_created: true
-- name: src/life/Life$LOCALE.class
-  visible: true
-  learner_created: true
-- name: src/life/Life$1.class
-  visible: true
-  learner_created: true
-- name: src/life/Life$STATE.class
-  visible: true
-  learner_created: true
-- name: src/test.class
-  visible: true
+  text: |
+    package life;
+
+    public interface View {
+        public void view();
+    }
   learner_created: true
 feedback_link: https://hyperskill.org/projects/54/stages/295/implement#comment
 status: Unchecked
