+++ Game of Life/task/src/life/engine/Model.java
@@ -1,15 +1,15 @@
-package life;
+package life.engine;
 
 public abstract class Model {
+
     int size;
-    int gen;
     protected int[][] map;
 
     Model(int size) {
         this.size = size;
         this.map = new int[size][size];
-        this.gen = 0;
     }
 
-    public void propagate(int gen) {}
+    public void propagate(int gens) {}
+    void generate() {}
 }
