+++ Game of Life/task/src/life/engine/View.java
@@ -1,4 +1,4 @@
-package life;
+package life.engine;
 
 public interface View {
     public void view();
