+++ course-remote-info.yaml
@@ -3,7 +3,7 @@ hyperskill_project:
   ide_files: https://stepik.org/media/attachments/lesson/209960/additional_files.json
   is_template_based: false
   use_ide: true
-update_date: Mon, 05 Jul 2021 03:53:42 UTC
+update_date: Mon, 05 Jul 2021 11:07:05 UTC
 stages:
 - id: 294
   step: 5436
